full stack
when the server starts, a master card list is generated from the newly included database (currently only including cards from beta). Then a random cube is generated (currently without any weights or concern about card rarity or distribution etc). Then 4 cards are taken from the cube and added to each of their respective 4 columns in the game state to start the draft. When a client loads the webpage from the server, it just tells the server they are the host which prompts the server to send the entire gamestate via websockets to the client as json data. The client gets names for each card that should be visible and goes to grab imgaes for each card from the web using the same url structure as curiosa.io, and displays them on the screen with bare minimum of formatting. Overall this represents the first pass all the way through the stack to display some cards on the user's screen.