connection screen
Now, when the host finishes configuring the draft, they are presented with a screen for awaiting the connection from the client. The screen provides a link to send to the client, and when the client follows the link, the 2 parties can participate in the draft, each selecting a column on their own turn, which updates the game state as well as the clients.

this represents the first time multi-user support is working in this app.

The server will likely crash right now if the cards run out, there's no graceful way to handle end of draft conditions, or anywhere for the app to go after the draft is complete yet.