inactive tabs
Before, when a tab was inactive, it wouldn't receive any state updates from the server, and would never get those missed updates, which could cause connected users to go out f sync with each other. I worked around this by making the page reload when it became active, which would at least synchronize the state at that point, and could have been accomplished without a full reload for something more elegant, but it would still miss network traffic while inactive, so the user wouldn't know that his partner had made a selection if he was on another window for example.

Now, when the page blurs I start a setInterval loop which for now sends a copy of the user connection packet to the server once per second, which seems to nudge chrome into being ok with receiving a websocket response. Then I clear the interval when the page regains focus and... it just works!

you can see changes initiated by your partner in real time even when the tab is not in focus.