ip hiding preliminary
currently the only way to connect to a host's server is to pass a link containing the host's IP address unless he uses some sort of URL forwarding or IP masking service.

I want people to feel comfortable streaming their draft sessions to viewers online, without feeling like they are exposing their IP address.

A number of changes have been put in place in preparation  to make changes that will mask the ip address and also make it more difficult for hosts to gain access to their connected client's id, preventing them from interfering with the client's draft process as the host.

So, we've done away with hashing the user ids in attempt to hide them in the links, it wasn't fully implemented in a way that was secure yet, but now that's being replaced.

The end goal is to use light encryption to create a long querystring parameter, too long to fit entirely on the screen in someone's url bar. The parameter will contain encrypted information about the host's ip address and any user's randomly generated id string in json format.

In the future, when either party connects to the server, it will be through a static github page associated with the project.  So when the host, client, or a spectator wants to sign into the site, all generated links will point to the github page, which will load the encrypted querystring parameter and obtain the ip address for the server, and then pull the necessary data, and use that to update the page contents via websockets, get requests, ajax calls, or other internal protocols that will prevent the navigated url or any generated links from ever containing the hosts ip address as they will always point to github, or potentially some other third party intermediary.

right now, the encrypted ip address is working as intended, but the intermediary site isnt up yet so the ip adrress is still included in links. Also the encryption key is not yet randomized.