random encryption key
now the key won't ever be 'secretkey', but will instead be random. The key is just stuck on the front of the encrypted querystring. Someone who want's to get the host IP from the querystring won't have too hard of a time doing it, if they have the link itself, but making the ip 100% irretrievable isnt the point. The point is to prevent the ip from leaking out in plaintext in the url bar or in shared links that might end up on the screen during streaming sessions. If a viewer does end up seeing the link, they'll have to type in several hundred characters and pass them through the encryption algorithm in order to reverse engineer the link. Not impossible but again thats not the point, and it should be enough of a deterrent to make server hosts and streamers comfortable.