session data json
now, host and user ids are not sent in raw text via querystring parameter anymore. The ip address of the host, along with the user ids are now added to json data which are padded with their own random data for both client and host, and then xored against their own random key for both data and host.  Now neither piece of information is readable in the url, and links can be passed safely. Of course the IP of the host is still included in the beginning of the url, but now that the data can all be encoded in the querystring, I can start to prepare to remove the ip from the links altogether.

also some javascript was moved from the index page to its own file in preparation to minimize the github document which will act as the intermediary to restore the ip information.