local testing
Previous updates which allowed for the github frontend also made it hard ot do local testing without new repo commits. This update adds a variable to sockethandler.py which lets the user switch into a local testing mode which pulls all of the scripts fromthe local server.

During testing, I learned that the certificate that i generated to make wss work, only works locally. The non-local github version works but gives a nasty security warning due to the self signed certificate.

Finally, the state won't update when the page is in focus, i need to extend the mthod that i used to monitor the websocket connection for network traffic when the page was out of focus.... to just do the same thing at all times whether in or out of focus