overhaul
Due to browser security restrictions, and a general lack of oversight on my part, I've decided to do a major overhaul on the project.

Working in the shadows, I've completely removed the python server for http and websockets. I've stopped attempting to circumvent security policy regarding mixed content. So no more self-signed ssl certificates and the like.

The entire project has been migrated to use a peer to peer approach using peer.js for webrtc communication.

So far I've got the main functionality ported, replicated and/or restored from the ashes of the original incarnation of the project, and with that accomplishment I can attempt to move forward with the deck management screen.