status message position
The status message wasn't showing in the right position on the host when the draft was starting up, now we run the method for repositioning elements around the deck manager window right after we add it and the status message moves to the right place.

This wasnt an issue on the client side.

Also, there's no longer a need for the method we were using to process card names from the old database, so thats been removed.