card ownership
Previously, all drafted cards went into a single array with a property added to each card object in the array to indicate which user selected (and therefore holds or owns) the card.

Now, there are 2 arrays, one for each user, when the guest selects a card, the message is sent to the host, who then knows that the guest owns the cards selected and updated the guest_cards array in the deck state, and then sends back a state update. When the host selects a column, he knows to update the host_cards array and then sends a state update.