configuration screen
Previously the default and only way to create a cube was to provide a number of cards, defaulted to 120, and random selection of cards equal to the number provided would be selected and added to the cube. There was no need for shuffling because the selection was already random. Duplicates were possible.

That mode has not been removed, and if no options are selected in the configuration screen, the exact same behavior will occur.

Now, however, there is an additional option, which is to provide a list of card names with optional associated quantity. The card names are one per line and entered into a text area. There is minimal error checking, the name needs to be exactly as it exists in the database or it won't work. Card quantities can be added as a number at the beginning of any line, and when the cube is constructed, it will include the correct quantity of each accurately named and numbered card. If the quantity is omitted, the cube generator will include exactly one of that card.

This was something that needed to be added eventually anyway, but the reason it was added now as to facilitate in the next planned feature, which is grouping of identical cards in players' drafted collections, followed by automatic sorting of cards. In order to make it much easier to obtain duplicate cards in a collection, the default cube selection under 'card name' mode is 40 rings of morrigan and 40 blinks.

The cube is also shuffled now before the draft.