duplicates
Previously, if a user had 2 of the same card in the same section of their collection, or otherwise, each card would show as a separate entry. Additionally, the only valid quantity was '1', so a user might have for example 6 copies of 1 quantity of blink all in their keep section.

Now, the array of cards in a users collection remains unaltered, but a new intermediary array is created from it , which currently is only used to consolidate identical cards in the same section and update their quantities before display. This also required keeping track of the last instance of a card in the original unaltered array, so that when a card is moved from one section of the collection to another, it always moves the last copy before recreating the altered array. This helps to preserve the order of the cards so that they don't jump around and shift order. Some order shifting still happens however, when the first card of a type is moved into a new section.

Sorting of cards by various criteria will be coming soon, and that should clear up any remaining jumping issues at which point i'll probably stop tracking the last instance of each card, as sorting of some kind will always be present, and that should eliminate any remaining position hopping.

On a similar note, I haven't needed to track the owner of a certain card explicitly since the collections were split into a separate list for the host and guest, so I also removed the property that was on each card indicating it's owner.