sort
Previously when cards were added to a user's collection after being selected in the draft, they would just be added to the end of the draft. Not long ago, there was no grouping of cards, or different sections within the collection for the cards to go into. In that environment it wasn't much of an issue to see the most recently drafted cards go to the bottom of the list.

Once the subjections of a user's collections were implemented, having cards arranged this way was causing some odd behaviors because... when you move a card from one subsection of your collection to another, the card still exists at the same index of the array that represents your collection. It's order hasn't changed at all, so if you move the card to another section, or back to its original section, it might appear to jump around to an unexpected order based on when the card was originally drafted rather than being based on when the card was moved. A simple solution to this might have been to track the subsections with individual arrays, but maintaining the draft order within your collection was never really a priority and instead it just pressed an inevitable feature that was going to need to be implemented eventually anyway, which is sorting.

After the grouping of cards was implemented in the last update, things got even stranger without sorting because the cards are still essentially ungrouped in the base array that represents your collection, and the grouping is really only happening at a cosmetic level, before being rendered into the deck management panel. Because of the way the grouping is being performed, combined with the way the collection is stored, i had to start keeping track of the index in the base array  of some instance of the card that was being grouped, so that it could be moved between subsections. When those cards were being moved around in the base array, they could affect the way the grouped cards were ordered on the screen within the subsections of the management panel.

Ultimately all of this made me feel it was time to get some sorting implemented. Right now, the framework is there to sort on any key in the card database, including name, mana cost, quantity, or any other metadata provided or that i decide to track in the future. However, I've just got something basic and functional in place right now that sorts the grouped cards alphabetically. In the future, the interface will be expanded to allow for a wide array of sorting options, but for now i want to press forward and drill all the way through the solution stack before getting too focused on finer details of interface features where possible. But the sorting really felt necessary to implement in some way now.

As a result of this I stopped tracking the last index a card occupies in the base array, and i just track the first time it's seen, since it all 'sorts out' afterward anyway.

Also, sometimes on the first page load, the browser would see that the main Draft object existed already, which would cause Main.js to look for the draft phase in order to determine which screen to display. This makes sense because the Draft object is created when Draft.js is loaded from index.html. But of course I didn't like the error so i decided to check for the existence of Draft.state.phase instead, which really should never exist when the page first loads.