Update DraftObject.js
Previously there was no end condition for the draft, and no way for the program to handle running out of cards in the cube. When you ran out of cards, the host would continue to try to pull them anyway, which would cause an error and that's it.

Now every time a card is about to be pulled from the cube, there is a check made to see if there are more than zero cards left in the cube to pull from.

The end result is that for the first time in this project, you can actually complete an entire draft, and the program remains responsive up to the very last column selection until the draft screen is empty... and beyond. The players can continue to sort their cards in their own collections indefinitely until the host closes the window.