Post deck screen
Now the project knows when the last card in the cube has been selected and therefore added to a user's collection, which triggers a change in the draft phase, to the new 'post_deck' phase, which will then trigger filling the #stage element with the new post deck screen. Currently the post deck screen is just a placeholder and that's coming next.